Get upload income page working.
