+++ src/settings/_index.njk
@@ -17,9 +17,6 @@
   </div>
   <div class="cluster small">
     <div>
-      <a href="/static/income-template.csv" download="income.csv">
-        Income CSV Template
-      </a>
       <a href="/static/savings-template.csv" download="savings.csv">
         Savings CSV Template
       </a>
@@ -28,12 +25,6 @@
       </a>
     </div>
   </div>
-
-  <div>
-    <label for="importIncomeFile" class="button">Upload Income</label>
-    <input id="importIncomeFile" type="file" name="import-income-file" class="visually-hidden" data-import-income-data>
-  </div>
-  <div data-import-income-progress aria-live="polite"></div>
 
   <div>
     <label for="importSavingsFile" class="button">Upload Savings</label>
