+++ src/settings/upload-income.njk
@@ -15,8 +15,20 @@
       Go Back
     </a>
   </div>
+
+  <div>
+    <a href="/static/income-template.csv" download="income.csv">
+      Income CSV Template
+    </a>
+  </div>
+
+  <div>
+    <label for="importIncomeFile" class="button">Upload Income</label>
+    <input id="importIncomeFile" type="file" name="import-income-file" class="visually-hidden" data-import-income-data>
+  </div>
+  <div data-import-income-progress aria-live="polite"></div>
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/settings/upload-income.js"></script>
+  <script src="/scripts/settings/upload-income.mjs" type="module"></script>
 {% endblock %}
